+++ app/src/main/java/com/teamwss/websoso/data/mapper/NovelMapper.kt
@@ -1,7 +1,9 @@
 package com.teamwss.websoso.data.mapper
 
 import com.teamwss.websoso.data.model.NovelDetailEntity
+import com.teamwss.websoso.data.model.NovelInfoEntity
 import com.teamwss.websoso.data.remote.response.NovelDetailResponseDto
+import com.teamwss.websoso.data.remote.response.NovelInfoResponseDto
 
 fun NovelDetailResponseDto.toData(): NovelDetailEntity {
     return NovelDetailEntity(
@@ -29,6 +31,31 @@ fun NovelDetailResponseDto.toData(): NovelDetailEntity {
             novelRating = novelRating,
             novelRatingCount = novelRatingCount,
             feedCount = feedCount,
+        ),
+    )
+}
+
+fun NovelInfoResponseDto.toData(): NovelInfoEntity {
+    return NovelInfoEntity(
+        novelDescription = novelDescription,
+        platforms = platforms.map {
+            NovelInfoEntity.PlatformEntity(
+                platformName = it.platformName,
+                platformImage = it.platformImage,
+                platformUrl = it.platformUrl,
+            )
+        },
+        attractivePoints = attractivePoints,
+        keywords = keywords.map {
+            NovelInfoEntity.KeywordEntity(
+                keywordName = it.keywordName,
+                keywordCount = it.keywordCount,
+            )
+        },
+        reviewCount = NovelInfoEntity.ReviewCountEntity(
+            watchingCount = watchingCount,
+            watchedCount = watchedCount,
+            quitCount = quitCount,
         ),
     )
 }
